change int to double for carbon values

+++ src/main/java/server/Action.java
@@ -7,8 +7,8 @@ public class Action {
     private String token;
     private String action;
     private int    value;
-    private int    carbonReduced;
-    private int    carbonProduced;
+    private double carbonReduced;
+    private double carbonProduced;
     private long   date;
     
     /**
@@ -16,10 +16,10 @@ public class Action {
      */
     public Action() {
     }
-
-    public Action(int carbonProduced, int carbonReduced){
-        this.carbonProduced=carbonProduced;
-        this.carbonReduced=carbonReduced;
+    
+    public Action(double carbonProduced, double carbonReduced) {
+        this.carbonProduced = carbonProduced;
+        this.carbonReduced = carbonReduced;
     }
     
     /**
@@ -42,8 +42,7 @@ public class Action {
      * @param carbonProduced carbon produced
      * @param carbonReduced  carbon saced
      */
-    public Action(String token, String action, int value, int carbonReduced, int carbonProduced) {
-        
+    public Action(String token, String action, int value, double carbonReduced, double carbonProduced) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -60,7 +59,7 @@ public class Action {
      * @param carbonReduced  carbon saced
      * @param date           date added (in miliseconds)
      */
-    public Action(String token, String action, int value, int carbonReduced, int carbonProduced, long date) {
+    public Action(String token, String action, int value, double carbonReduced, double carbonProduced, long date) {
         this.token = token;
         this.action = action;
         this.value = value;
@@ -77,7 +76,7 @@ public class Action {
      * @param carbonReduced  carbon saced
      * @param date           date added (in miliseconds)
      */
-    public Action(String action, int value, int carbonReduced, int carbonProduced, long date) {
+    public Action(String action, int value, double carbonReduced, double carbonProduced, long date) {
         this.action = action;
         this.value = value;
         this.carbonReduced = carbonReduced;
@@ -134,14 +133,14 @@ public class Action {
      * getter for carbonReduced.
      * @return carbonReduced
      */
-    public int getCarbonReduced() {
+    public double getCarbonReduced() {
         return carbonReduced;
     }
     
     /**
      * setter for carbonReduced.
      */
-    public void setCarbonReduced(int carbonReduced) {
+    public void setCarbonReduced(double carbonReduced) {
         this.carbonReduced = carbonReduced;
     }
     
@@ -149,14 +148,14 @@ public class Action {
      * getter for Value.
      * @return carbonProduced
      */
-    public int getCarbonProduced() {
+    public double getCarbonProduced() {
         return carbonProduced;
     }
     
     /**
      * setter for Value.
      */
-    public void setCarbonProduced(int carbonProduced) {
+    public void setCarbonProduced(double carbonProduced) {
         this.carbonProduced = carbonProduced;
     }
